refactor phase 2: added StylesPreviewPresentation
