+++ PowerPointLabs/PowerPointLabs/ImageSearch/ImageSearchPane.xaml
@@ -12,7 +12,8 @@
              Height="600"
              WindowTitleBrush="#D74926"
              Closing="ImageSearchPane_OnClosing"
-             GlowBrush="#CACACA"
+             BorderThickness="1"
+             BorderBrush="#D74926"
              IsKeyboardFocusWithinChanged="ImageSearchPane_OnIsKeyboardFocusWithinChanged">
     <controls:MetroWindow.Resources>
         <ResourceDictionary>
@@ -31,7 +32,7 @@
             <ColumnDefinition Width="auto"></ColumnDefinition>
             <ColumnDefinition Width="*"></ColumnDefinition>
         </Grid.ColumnDefinitions>
-        <GridSplitter Name="Splitter" DragDelta="Splitter_OnDragDelta" Grid.Column="1" ResizeDirection="Columns" Width="3" Height="Auto" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" />
+        <GridSplitter Name="Splitter" Margin="2,0,0,0" DragDelta="Splitter_OnDragDelta" Grid.Column="1" ResizeDirection="Columns" Width="3" Height="Auto" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" />
         <Grid Name="StylesPreviewGrid" Margin="10,0,0,0" Grid.Column="0">
             <Grid.RowDefinitions>
                 <RowDefinition Height="5" />
