+++ PowerPointLabs/PowerPointLabs/ImageSearch/ImageSearchPane.xaml.cs
@@ -13,8 +13,10 @@ using MahApps.Metro.Controls.Dialogs;
 using Microsoft.Office.Core;
 using PowerPointLabs.AutoUpdate;
 using PowerPointLabs.ImageSearch.Model;
+using PowerPointLabs.ImageSearch.Presentation;
 using PowerPointLabs.ImageSearch.SearchEngine;
 using PowerPointLabs.ImageSearch.SearchEngine.Options;
+using PowerPointLabs.ImageSearch.SearchEngine.VO;
 using PowerPointLabs.ImageSearch.Util;
 using PowerPointLabs.Models;
 using PowerPoint = Microsoft.Office.Interop.PowerPoint;
@@ -37,9 +39,8 @@ namespace PowerPointLabs.ImageSearch
 
         private const int TimerInterval = 2000;
 
-        // TODO REFACTOR customize presentation that contains preview processing logic
         // a background presentation that will do the preview processing
-        public PowerPointPresentation PreviewPresentation { get; set; }
+        public StylesPreviewPresentation PreviewPresentation { get; set; }
 
         public GoogleEngine SearchEngine { get; set; }
 
@@ -70,71 +71,29 @@ namespace PowerPointLabs.ImageSearch
         {
             // TODO MUST load options from config
             SearchEngine = new GoogleEngine(new GoogleOptions())
-            .WhenSucceed((searchResults, startIdx) =>
-            {
-                if (searchResults.Items == null)
-                {
-                    for (var i = startIdx + 0; i < startIdx + GoogleEngine.NumOfItemsPerRequest; i++)
+                .WhenSucceed(WhenSearchSucceed())
+                .WhenCompleted(WhenSearchCompleted())
+                .WhenFail(response => {
+                    Dispatcher.BeginInvoke(new Action(() =>
                     {
-                        SearchList[i].IsToDelete = true;
-                    }
-                    return;
-                }
-                for (var i = 0; i < GoogleEngine.NumOfItemsPerRequest; i++)
-                {
-                    var item = SearchList[startIdx + i];
-                    if (i < searchResults.Items.Count)
-                    {
-                        var searchResult = searchResults.Items[i];
-                        var targetLocation = TempPath.GetPath("thumbnail");
+                        this.ShowMessageAsync("Error",
+                            "Failed to search images. Please check your network, or the daily API quota is ran out.");
+                    }));
+                });
+        }
 
-                        // intent: 
-                        // download thumbnail and show it,
-                        // also save other meta info (e.g. full-size img link)
-                        new Downloader()
-                            .Get(searchResult.Image.ThumbnailLink, targetLocation)
-                            .After(() =>
-                            {
-                                item.ImageFile = targetLocation;
-                                item.FullSizeImageUri = searchResult.Link;
-                            })
-                            .Start();
-                    }
-                    else
-                    {
-                        item.IsToDelete = true;
-                    }
-                }
-            })
-            .WhenFail(response =>
-            {
-                Dispatcher.BeginInvoke(new Action(() =>
-                {
-                    this.ShowMessageAsync("Error",
-                        "Failed to search images. Please check your network, or the daily API quota is ran out.");
-                }));
-            })
-            .WhenCompleted(isSuccess =>
+        private GoogleEngine.WhenCompletedEventDelegate WhenSearchCompleted()
+        {
+            return isSuccess =>
             {
                 Dispatcher.BeginInvoke(new Action(() =>
                 {
                     SearchProgressRing.IsActive = false;
-                    var isNoMoreSearchResults = false;
-                    for (int i = 0; i < SearchList.Count;)
-                    {
-                        if (SearchList[i].IsToDelete)
-                        {
-                            SearchList.RemoveAt(i);
-                            isNoMoreSearchResults = true;
-                        }
-                        else
-                        {
-                            i++;
-                        }
-                    }
+                    var isThereMoreSearchResults = !RemoveElementInSearchList(item => item.IsToDelete);
                     if (isSuccess 
-                        && !isNoMoreSearchResults 
-                        && SearchList.Count + GoogleEngine.NumOfItemsPerRequest - 1/*load more item*/ <= GoogleEngine.MaxNumOfItems)
+                        && isThereMoreSearchResults 
+                        && SearchList.Count + GoogleEngine.NumOfItemsPerRequest - 1/*loadMore item*/ 
+                        <= GoogleEngine.MaxNumOfItems)
                     {
                         SearchList.Add(new ImageItem
                         {
@@ -142,12 +101,60 @@ namespace PowerPointLabs.ImageSearch
                         });
                     }
                 }));
-            });
+            };
+        }
+
+        // TODO util
+        private bool RemoveElementInSearchList(Func<ImageItem, bool> cond)
+        {
+            var isAnyElementRemoved = false;
+            for (var i = 0; i < SearchList.Count;)
+            {
+                if (cond(SearchList[i]))
+                {
+                    SearchList.RemoveAt(i);
+                    isAnyElementRemoved = true;
+                }
+                else
+                {
+                    i++;
+                }
+            }
+            return isAnyElementRemoved;
+        }
+
+        private GoogleEngine.WhenSucceedEventDelegate WhenSearchSucceed()
+        {
+            return (searchResults, startIdx) =>
+            {
+                searchResults.Items = searchResults.Items ?? new List<SearchResult>();
+                for (var i = 0; i < GoogleEngine.NumOfItemsPerRequest; i++)
+                {
+                    var item = SearchList[startIdx + i];
+                    if (i >= searchResults.Items.Count)
+                    {
+                        item.IsToDelete = true;
+                        continue;
+                    }
+
+                    var searchResult = searchResults.Items[i];
+                    var thumbnailPath = TempPath.GetPath("thumbnail");
+
+                    new Downloader()
+                        .Get(searchResult.Image.ThumbnailLink, thumbnailPath)
+                        .After(() =>
+                        {
+                            item.ImageFile = thumbnailPath;
+                            item.FullSizeImageUri = searchResult.Link;
+                        })
+                        .Start();
+                }
+            };
         }
 
         private void InitPreviewPresentation()
         {
-            PreviewPresentation = new PowerPointPresentation(TempPath.TempFolder, "ImagesLabPreview");
+            PreviewPresentation = new StylesPreviewPresentation();
             PreviewPresentation.Open(withWindow: false, focus: false);
         }
 
@@ -300,249 +307,20 @@ namespace PowerPointLabs.ImageSearch
                 var selectedId = PreviewListBox.SelectedIndex;
                 PreviewProgressRing.IsActive = true;
                 PreviewList.Clear();
+                PreviewPresentation.PreviewStyles(imageItem);
 
-                var previewFile = TempPath.GetPath("original"); // --> opt
-                var previewFile2 = TempPath.GetPath("directText");
-                var previewFile3 = TempPath.GetPath("overlay"); // --> opt for direct text
-                var previewFile4 = TempPath.GetPath("textbox"); // --> opt for blur textbox
-                var previewFile5 = TempPath.GetPath("blur");
-                var previewFile6 = TempPath.GetPath("blur_textbox");
-                var previewFile7 = TempPath.GetPath("blur_part"); // --> opt for blur
-
-                // TODO DRY
-                PowerPointSlide thisSlide;
-                try
-                {
-                    thisSlide = PreviewPresentation.AddSlide(PowerPointCurrentPresentationInfo.CurrentSlide.Layout);
-                }
-                catch
-                {
-                    thisSlide = PreviewPresentation.AddSlide();
-                }
-                thisSlide.DeleteAllShapes();
-                try
-                {
-                    PowerPointCurrentPresentationInfo.CurrentSlide.Shapes.Range().Copy();
-                    thisSlide.Shapes.Paste();
-                    foreach (PowerPoint.Shape shape in thisSlide.Shapes)
-                    {
-                        if (shape.Name.StartsWith("pptImagesLab"))
-                        {
-                            shape.Delete();
-                        }
-                    }
-                }
-                catch
-                {
-                    // nothing to copy-paste
-                }
-                var imageShape = thisSlide.Shapes.AddPicture(imageItem.FullSizeImageFile ?? imageItem.ImageFile, 
-                    MsoTriState.msoFalse, MsoTriState.msoTrue, 0,
-                    0);
-                FitToSlide.AutoFit(imageShape, PreviewPresentation);
-                imageShape.ZOrder(MsoZOrderCmd.msoSendToBack);
-
-//                thisSlide.GetNativeSlide().Export(previewFile, "JPG");
-//                PreviewList.Add(new ImageItem
-//                {
-//                    ImageFile = previewFile
-//                });
-                // Original Preview done here
-
-                // Textbox style 1 starts
-                foreach (PowerPoint.Shape shape in thisSlide.Shapes)
-                {
-                    if (shape.Type == MsoShapeType.msoPlaceholder
-                        || shape.Type == MsoShapeType.msoTextBox)
-                    {
-                        if (shape.TextFrame.HasText == MsoTriState.msoFalse
-                            || shape.Tags["GotHighlighted"].Trim().Length != 0)
-                        {
-                            continue;
-                        }
-
-                        // filled by added shape (can control size)
-                        shape.Fill.Visible = MsoTriState.msoFalse;
-                        shape.Line.Visible = MsoTriState.msoFalse;
-
-                        var whiteColor = Color.White;// TODO customize
-                        var r = whiteColor.R;
-                        var g = whiteColor.G;
-                        var b = whiteColor.B;
-
-                        var rgb = (b << 16) | (g << 8) | (r);
-                        var font = shape.TextFrame2.TextRange.TrimText().Font;
-                        font.Fill.ForeColor.RGB = rgb;
-                        font.Bold = MsoTriState.msoFalse;
-                        font.Name = "Segoe UI Light"; // TODO customize
-
-                        var textEffect = shape.TextEffect;
-                        textEffect.FontSize += 10;
-                    }
-                }
-                thisSlide.GetNativeSlide().Export(previewFile2, "JPG");
                 PreviewList.Add(new ImageItem
                 {
-                    ImageFile = previewFile2
+                    ImageFile = PreviewPresentation.DirectTextStyleImagePath
                 });
-                // Textbox style 1 ends
-//                // Textbox style 2 starts
-//                var overlayShape = thisSlide.Shapes.AddShape(MsoAutoShapeType.msoShapeRectangle,
-//                                                                0,
-//                                                                0,
-//                                                                PowerPointPresentation.Current.SlideWidth,
-//                                                                PowerPointPresentation.Current.SlideHeight);
-//                overlayShape.Fill.ForeColor.RGB = Utils.Graphics.ConvertColorToRgb(Color.Black);
-//                overlayShape.Fill.Transparency = 0.65f;
-//                overlayShape.Line.Visible = MsoTriState.msoFalse;
-//                overlayShape.ZOrder(MsoZOrderCmd.msoSendToBack);
-//                imageShape.ZOrder(MsoZOrderCmd.msoSendToBack);
-//                thisSlide.GetNativeSlide().Export(previewFile3, "JPG");
-//                PreviewList.Add(new ImageItem
-//                {
-//                    ImageFile = previewFile3
-//                });
-//                overlayShape.Delete();
-                //
-                // textbox style 5 starts
-                var overlayShape = thisSlide.Shapes.AddShape(MsoAutoShapeType.msoShapeRectangle,
-                                                                0,
-                                                                0,
-                                                                PowerPointPresentation.Current.SlideWidth,
-                                                                PowerPointPresentation.Current.SlideHeight);
-                overlayShape.Fill.ForeColor.RGB = Utils.Graphics.ConvertColorToRgb(Color.Black);
-                overlayShape.Fill.Transparency = 0.85f;
-                overlayShape.Line.Visible = MsoTriState.msoFalse;
-
-                if (imageItem.BlurImageFile == null)
-                {
-                    var blurImageFile = TempPath.GetPath("fullsize_blur");
-                    using (var imageFactory = new ImageFactory())
-                    {
-                        var image = imageFactory.Load(imageItem.ImageFile);
-                        image = image.GaussianBlur(5);
-                        image.Save(blurImageFile);
-                        if (image.MimeType == "image/png")
-                        {
-                            blurImageFile += ".png";
-                        }
-                        imageItem.BlurImageFile = blurImageFile;
-                    }
-                }
-                var blurImageShape = thisSlide.Shapes.AddPicture(imageItem.BlurImageFile,
-                    MsoTriState.msoFalse, MsoTriState.msoTrue, 0,
-                    0);
-                FitToSlide.AutoFit(blurImageShape, PreviewPresentation);
-                overlayShape.ZOrder(MsoZOrderCmd.msoSendToBack);
-                blurImageShape.ZOrder(MsoZOrderCmd.msoSendToBack);
-                imageShape.ZOrder(MsoZOrderCmd.msoSendToBack);
-
-                thisSlide.GetNativeSlide().Export(previewFile5, "JPG");
                 PreviewList.Add(new ImageItem
                 {
-                    ImageFile = previewFile5
+                    ImageFile = PreviewPresentation.BlurStyleImagePath
                 });
-
-                overlayShape.ZOrder(MsoZOrderCmd.msoSendToBack);
-
-                // blur textbox region starts
-                var listOfBlurImageCopy = new List<PowerPoint.Shape>();
-                foreach (PowerPoint.Shape shape in thisSlide.Shapes)
-                {
-                    if (shape.Type == MsoShapeType.msoPlaceholder
-                        || shape.Type == MsoShapeType.msoTextBox)
-                    {
-                        if (shape.TextFrame.HasText == MsoTriState.msoFalse
-                            || shape.Tags["GotBlured"].Trim().Length != 0)
-                        {
-                            continue;
-                        }
-                        // multiple paragraphs.. 
-                        foreach (TextRange2 paragraph in shape.TextFrame2.TextRange.Paragraphs)
-                        {
-                            if (paragraph.TrimText().Length > 0)
-                            {
-                                blurImageShape.Copy();
-                                var blurImageShapeCopy = thisSlide.Shapes.Paste()[1];
-                                listOfBlurImageCopy.Add(blurImageShapeCopy);
-                                PowerPointLabsGlobals.CopyShapePosition(blurImageShape, ref blurImageShapeCopy);
-                                blurImageShapeCopy.PictureFormat.Crop.ShapeLeft = paragraph.BoundLeft - 5;
-                                blurImageShapeCopy.PictureFormat.Crop.ShapeWidth = paragraph.BoundWidth + 10;
-                                blurImageShapeCopy.PictureFormat.Crop.ShapeTop = paragraph.BoundTop - 5;
-                                blurImageShapeCopy.PictureFormat.Crop.ShapeHeight = paragraph.BoundHeight + 10;
-                                var overlayBlurShape = thisSlide.Shapes.AddShape(MsoAutoShapeType.msoShapeRectangle,
-                                                                paragraph.BoundLeft - 5,
-                                                                paragraph.BoundTop - 5,
-                                                                paragraph.BoundWidth + 10,
-                                                                paragraph.BoundHeight + 10);
-                                overlayBlurShape.Fill.ForeColor.RGB = Utils.Graphics.ConvertColorToRgb(Color.Black);
-                                overlayBlurShape.Fill.Transparency = 0.85f;
-                                overlayBlurShape.Line.Visible = MsoTriState.msoFalse;
-                                listOfBlurImageCopy.Add(overlayBlurShape);
-                                Utils.Graphics.MoveZToJustBehind(blurImageShapeCopy, shape);
-                                Utils.Graphics.MoveZToJustBehind(overlayBlurShape, shape);
-                                shape.Tags.Add("GotBlured", blurImageShapeCopy.Name);
-                            }
-                        }
-                    }
-                }
-                blurImageShape.ZOrder(MsoZOrderCmd.msoSendToBack);
-
-                thisSlide.GetNativeSlide().Export(previewFile6, "JPG");
                 PreviewList.Add(new ImageItem
                 {
-                    ImageFile = previewFile6
+                    ImageFile = PreviewPresentation.TextboxStyleImagePath
                 });
-
-                foreach (var shape in listOfBlurImageCopy)
-                {
-                    shape.Delete();
-                }
-                
-                blurImageShape.Delete();
-                overlayShape.Delete();
-
-                //
-                // Textbox style 3 starts
-//                foreach (PowerPoint.Shape shape in thisSlide.Shapes)
-//                {
-//                    if (shape.Type == MsoShapeType.msoPlaceholder
-//                        || shape.Type == MsoShapeType.msoTextBox)
-//                    {
-//                        if (shape.TextFrame.HasText == MsoTriState.msoFalse
-//                            || shape.Tags["GotHighlighted"].Trim().Length != 0)
-//                        {
-//                            continue;
-//                        }
-//                        // multiple paragraphs.. 
-//                        foreach (TextRange2 paragraph in shape.TextFrame2.TextRange.Paragraphs)
-//                        {
-//                            if (paragraph.TrimText().Length > 0)
-//                            {
-//                                var highlightShape = thisSlide.Shapes.AddShape(MsoAutoShapeType.msoShapeRectangle,
-//                                                                paragraph.BoundLeft - 5,
-//                                                                paragraph.BoundTop - 5,
-//                                                                paragraph.BoundWidth + 10,
-//                                                                paragraph.BoundHeight + 10);
-//                                highlightShape.Fill.ForeColor.RGB = Utils.Graphics.ConvertColorToRgb(Color.Black); // TODO customize
-//                                highlightShape.Line.Visible = MsoTriState.msoFalse;
-//                                Utils.Graphics.MoveZToJustBehind(highlightShape, shape);
-//                                highlightShape.Name = "PPTLabsHighlightBackgroundShape" + DateTime.Now.ToString("yyyyMMddHHmmssffff");
-//                                highlightShape.Tags.Add("HighlightBackground", shape.Name);
-//                                shape.Tags.Add("GotHighlighted", highlightShape.Name);
-//                            }
-//                        }
-//                    }
-//                }
-//                thisSlide.GetNativeSlide().Export(previewFile4, "JPG");
-//                PreviewList.Add(new ImageItem
-//                {
-//                    ImageFile = previewFile4
-//                });
-
-                //
-                // dont affect next time preview
-                thisSlide.Delete();
                 PreviewListBox.SelectedIndex = selectedId;
 
                 // try catch finally?
